Attach to RoboChart metamodel and grammar
As if by magic, this means that any rcert in a RoboChart project now
properly checks the references for RoboChar

The attachment to the RoboChart grammar could be dropped if we
reimplement QualifiedName (and, eventually, the expression grammar).
I'm not sure which is worse: doing that, or pulling in the RoboChart
textual plugins as dependencies.

This required a _lot_ of scratching around and copying of MWE2 and other
such things from existing projects to work out how to get this to work.
Hopefully I'll remember how to do this in future, or at least remember
this commit exists.

Note to self: make sure the genmodel does NOT generate its own copy of
RoboChart.  This causes issues!