Add a workaround for importing platform files
This is just a workaround for now, since it pulls in any coreassertions
file that contains a module referenced in a module target.  It won't
work properly if we have multiple modules referenced in a file.

The idea is that eventually we won't use coreassertions at all, and can
move to importing the 'defs/' directly, in the same way that the
existing assertions language does.