Fallout from trying to use Sirius
I'm not actually sure how many of these changes are necessary, but there
was a load of rearranging imports and suchlike when I was trying to get
Sirius to understand the relationship between robocalc and robochart,
and this is just the aftermath.

Not sure how 'RoboCert' became 'Robocert', but the name is provisional
anyway.