Add allow/forbid sets to 'anything..until'
These gaps (now called 'loose sequence gaps') now have the concrete
syntax 'anything [in SET] [except SET] until', where SET is a brace
delimited enumeration of arrow events (eg 'operation X() from Y to Z').
It's not clear to me whether these set members need to be something
other than arrow events for now.

The semantics is a little inelegant for allow sets: an empty or
nonexistent allow set is considered to be the universe set.  This is
to simplify the metamodel and because right now I can't think of a use
for an empty allow set (whereas there is an obvious use for an empty
forbid set).

Fixes #1.