Overhaul arrows in metamodel
This change introduces 'message specs' as the replacement for the vague
and overloaded 'arrow' concept.  Anything like 'operation X from Y to Z'
is a message spec, which may or may not be part of an arrow.  Gap
message sets now track these instead of arrow actions (which are now
wrappers over message specs).

There is a class distinction between 'arrow' message specs and 'gap'
message specs, but I'm not sure if I'll keep it.