Use /\ for loose gaps.
The semantics for steps now treats gaps and actions mostly separately,
with the exception that the gap exclude set now needs to include the
CSP events on which the action will communicate.

All arrow actions now have the form (event -> SKIP), which is nice and
uniform.

Final actions are translated to STOP for now.  I need to work out how
to handle a termination requirement.

Fixes #6.
Fixes #7.