Instantiate constants directly
This'll move us off needing to target the coreassertions file.

An open question is whether we instantiate to the values in
instantiations.csp, or to the type defaults.  I've put the type defaults
in for now, but I might revisit this later.

Another one is what to do with the growing collection of extensions I've
cribbed from the CSP generator but refactored a bit...