Add target overriding on assertions
For now, the syntax is just 'ASSERTION on TARGET', where TARGET can be
the override target form added previously.  This completely replaces the
sequence's standard target - maybe one day we'll want to override it
in place, at which point we'll probably need a more complex setup to
install an override target on top of the sequence's override target.
(Perhaps a new target that is literally 'target of this sequence'?)

Closes #17.