Overhaul gaps
There is now only one kind of gap - no more strict and loose gaps.
Instead, there are now two types of gap message set: extensional
(as before), and universe.  Allow sets may be either; forbid sets are
always extensional.

When no specific gap stanza is written, the result is as if one
wrote 'anything in {} except {}'.  There is currently no specific
syntax for declaring that one wants the universe allow set: it's assumed
when someone types 'anything' but not 'in {...}'.

Report needs updating.