Add initial argument support to message specs
Both gap and arrow specs can now have expressions as arguments, or the
'...' (rest) argument, which tells the generator to pad with ?_ in
prefixes and leave open in event sets.  (This argument may go away later
on, as it's a fairly low level window into what works well for CSPM.)

Other arguments will become available as and when case studies need
them.

Report now out of sync again.