Make arrow direction explicit and actors implicit
Now that we use derived members of EMF objects liberally, I've decided
to switch the metamodel around so that the from/to/target references of
messages are derived (by walking up to the sequence group) and the
direction is explicit.  This both simplifies a lot of the generation
code (since the direction is the main thing the generator needs) and
paves the way for eliminating actor naming, which simplifies the textual
syntax.

On the Sirius end, we should be able to use the derived references to
keep things moving.

Syntax is provisional.

Report needs amending.