Report: reorganise semantics
Well-formedness is now going to be in the individual notation parts, whereas
I'm (for now) keeping the semantics to its own part.  This mostly recognises
the plugin layout: the semantics is a generator, which is a separate concern to
the language definitions.  Maybe if I was being a purist I'd split the syntax
sections into their own part too, hmm.