Make loop names optional
They aren't used for anything yet, and indeed in many diagrams won't
be used for anything.  They'll only really be used for break
statements, which haven't landed.

I've added some syntactic sugar so that the ':' between name and bound
is optional for nameless loops.  Honestly, I'm unsure about this, it
might be better to exclude ':' here.