Make groups a top-level RAPackage unit
I might live to regret this, since I'm not 100% whether the idea
of groups makes sense for CSP yet.  But for things like sequences
and hypothetical other notations, it makes sense to have fenced
regions of the RAPackage that delimit a particular context,
and it makes sense for the metamodel to present that structure.