Report: start work on low-level notation chapters
I've decided these are probably more a separate pseudo-notation than
a core part of the language, so I've started moving them into a separate
position.  This also makes it easier to cut them out of the language
later on, if we want to make them a separate 'expert mode' in robocert.