Mostly implement tick-tock refinement
We now have a more or less working CSP property generator, with a few
teething issues that are probably worth separating out into commits,
and have recast the sequence property generator as lifting into CSP.

This commit has broken the implementation of instantiation on the
assertion level, but instantiation needs reworking anyway (#53 for
instance) and so I'll work on this soon.

Fixes #36.

Fixes #42.