Use handrolled expression language
I was having issues trying to work out how to get constants to work with
the RoboChart expression language, and expected I'd have even more
problems working out how to get bound parameters to work with the
RoboChart expression language.  So, for now, I've scrapped it and
replaced it with an extremely simple RoboCert one.

The RoboCert language supports only two things right now: integer
literals and constants.  Eventually it'll have parameter bindings,
and maybe some form of arithmetic, but I'll quickly run into the
problem of typing if I get to that stage.

Fixes #25.