Make all actions have 'anytime' semantics
In other words, actions can now always nondeterministically choose to
wait a tock before firing.

It's unclear as to whether this is the right semantics, but it seems to
be less wrong than what we previously had.  See discussion with Pedro
and contemporary issues (none of which I'm closing yet, just in case).