Always emit gaps, even when inactive
This papers over a semantic wrinkle where TCHAOS({}) /\ P is not
equivalent to P (since the former can timelock).  It also papers over
a semantic question of whether sequences should permit models to impose
deadlines, when, and how... likely to be discussed at length later.

The exact definition of how a gap wraps a process is now a lot more
opaque, to make further changes a little easier.