Fix lack of partiality in tick-tock
The intended semantics of a sequence is that, unless otherwise
specified, the sequence is partial in its specification of behaviour
and open at the end.  That is, the model can engage in any behaviour it
desires after the end of the sequence.

This worked in trace semantics, but not in tick-tock semantics,
because the final part of any CSP elaboration of a sequence that didn't
otherwise finish in an infinite loop was SKIP, and this requires
termination.  What we instead need at the end of a sequence is chaos,
so I've added this to the outer-level sequence generator.

With this change, we also need a way of making sure 'anything in X
until end' has the appropriate semantics.  This required a further
change to the semantics of FinalAction to timed STOP, which, as far as
I know, leaves the effective semantics of such a step as 'TCHAOS(X)'.

We'll have to think more about how to capture required termination,
what to name FinalAction, etc etc.

Fixes #66.