Re-rename WildcardArguments
I realised while writing the reference manual section for what were
then BindingArguments that calling BindingArguments BindingArguments,
when they might not actually bind anything, sounds weird.  So I've
reverted this change.

Eventually we might want to generalise the notion of something being
a binding spot, and we might be back to square one.