Make RoboCert whitespace-sensitive
This is primarily to make sequences more compact, less cluttered with
symbols, and easier to read.  (And also because I didn't previously
know that this was an option, and it seemed like a good solution to my
long-running debate as to how to delimit blocks/subsequences/etc).

RoboCert now looks a lot more like Python, for better or for worse.