Make steps towards generating memories
We now have some code that determines what the bindings are in a
sequence, and generates a module full of channel definitions for each
sequence that has such bindings.

I forgot to mention in the last message that I'm experimenting with
Java 17, and migrating off Xtend in some places.  This is because of the
continuing uncertainty as to the future of Xtend, and also to try rein
in my tendency to use Xtend magic all over the place.  I might reverse
it.