Add minus-exprs, accidentally Java-ify a lot of things
I might backtrack on the Java conversion eventually.  I did it because
I thought that changing one or two stubs to Java would be a decent
trade-off, and then _all_ of the stubs automatically turned back into
Java and I'm not so sure.

Conformance has changed to Java 17, as part of a failed attempt to get
switch patterns working.  They're preview in 17, and even if
I enable preview functionality in my packages, Eclipse can't load the
result into its UI.