Allow parsing and generation of empty subsequences
An issue with significant-whitespace syntax is that we need a particular
marker to say that a subsequence is empty, so I've added 'nothing'.

It then transpired that the generator wasn't properly handling this case
(it was emitting an empty string instead of SKIP), so I've fixed that.

Why are empty subsequences needed now?  As the 'else' leg of optional or
alternative blocks.

Parsing tests for subsequences are on their way.