Implement first pass of actor pairs
There are now three ways to state which actors form the ends of a
message:

- explicit (both ends by name, 'from X to Y')
- directional ('in' or 'out')
- implicit (only works with operations, equivalent to 'out')

Scoping and the ability to reference worlds by name are not yet working,
so ExplicitActorPairs aren't quite finished.  That said, directional
and implicit pairs seem to be working fine (FDR4 crashed when I tried
them, but FDR4 is wont to crash randomly, and the unit tests all point
in the right direction.  Will check.)