Implement initial core assertions
These are: determinism, timed deadlock, timelock, and termination.

The particular implementation is a little bit ugly at the moment, as a
weird mismatch of different Java patterns for constructing code, but I'm
sure I'll settle down on something useful.

As discussed with Pedro, clock initialisation and state reachability
are not a priority right now.

Fixes #37.