[wip] Part 1 of actor overhaul
What I'm trying to do here is make it so that SequenceGroups have one
specific 'target' that is named, but permit any number of Actors that
either reference relationships to the target (eg 'target', 'world') or,
eventually, specific subcomponents of the target.  These two types of
Actor will lead to two different sorts of sequence diagram.

This commit goes most of the way to adding this, but is a WIP because
I've found out that the particular approach I'm using here (having each
Sequence name the actors it wants to use) doesn't work.  Message sets
are rooted at the SequenceGroup level and do not have access to a
Sequence from which they can get outputs!

So... the next commit will move actors back to SequenceGroups.  And then
we have the decision of whether to make the split between the two sorts
of group (target/world, and target/subcomponents) rigid in the
metamodel, or keep it for validation.