Start overhauling target use
Because Targets are now just generally 'the target of any specification'
we now have the idea of a SpecGroup that captures a group of
specifications targeting a Target.  CSPGroups have become SpecGroups,
but the transformation to something that properly understands target
parameterisation is not quite there yet.

Next step, I think, is to make Targets inhabit their own Group, which
will allow the common boilerplate such as message sets and tick-tock
contexts to be centralised.