Get target overhaul to steady state
Whew, this took all day and I'm still feeling like the end result is
worse than it was when I started.

What I was trying to do was to make it so that targets are specified in
a new form of group, with all forms of specification referencing those
targets, the tick-tock contexts being generated inside the target
groups, etc.  We're still not all the way there yet, because, for
instance, CSP processes can't pick up their target's tick-tock module,
nor can they easily define a new target by themselves.  Hmm.