Simplify actor pair metamodel
Actor pairs are now 'edge's, and there is no longer a distinction
between directional and implicit; the former has been renamed to the
latter and we just assume an outbound direction if none has been named.

Added new validation rules for edges make sure that edges can't go
backwards or loop or anything like that.