Lay groundwork for multiple lifelines
We're going to try something similar to Lucas Lima et al.'s 2014
encoding of UML sequence diagrams into Compass CML, which involves
having an alphabetised outer parallelisation.

So far, this just adds the notion of going through the lifetime list and
preparing some contexts to be (eventually) threaded through the
semantics.  There is an optimisation for the 1-lifeline case that
just delegates back to the existing semantics, so we don't do anything
different yet.