Migrate ImportGenerator to Java/streams
Part of a general push for me to try to make things consistent and
stream-based, and to think again about infrastructural decisions I made
early on.

Some parts of ImportGenerator have been replaced with calls to
CTimedGeneratorUtils; I might push this further, as I don't want to
duplicate that too much.