De-Xtend and clean up more unit tests
In honesty, most of this came from chasing a goal of getting what I have
at the moment building and operational in IntelliJ, as I find that a
much more comfortable IDE to work in (except of course where I need to
test the Eclipse specific parts!)

I haven't committed the IntelliJ files yet; I'm not sure if I will, as
they're quite idiosyncratic to how I'm trying to use the IDE, and also
use a lot of working-around the fact that this is an Eclipse project
first and foremost.

This somehow took all day.  I'm annoyed at this, as today was supposed
to be about threading through lifeline context more.