Rewrite import generator
This is me trying to chase down a bug that, in the end, had little to
nothing to do with the bits that got rewritten.

I've left in place a workaround that omits an anonymity check in the
import generator, as certain imports aren't appearing where they should
with it.  I'll ask around to see why this is necessary.