Fix broken tests on Eclipse, hopefully
This was because I'd switched to using Hamcrest 2 when developing the
tests in IDEA, but forgot to tell Eclipse to use it; instead, Eclipse
picked up Hamcrest 1 as a transitive dependency of Xtext testing and
this isn't forwards compatible.  The plugin now explicitly requires
Hamcrest 2.

Hamcrest 2 is available in the most recent Eclipse Orbit repositories.
Remind me to write a readme note about this.