Make UntilFragments a form of BlockStep
This is an experimental change to the metamodel (while we're finalising
the metamodel for version 0) that effectively makes UntilFragments a
combined fragment.  The idea is now that any sequence of fragments can
serve as the release for the until condition, so long as there is a
well defined notion of an 'initial set' for the sequence (this point to
discuss with Ana).

Justification is that we can now have constructs of the form 'anything
until either A or B happens', or 'anything until this loop begins',
or even possibly just 'anything until nothing' eg. loop forever (though
this is maybe equivalent to anything until end'.

There will be well-formedness conditions needing to be added for this.