Tidy up occurrences in metamodel
We've now committed to the erstwhile FinalAction being deadlock.
Termination may be modelled later on, if we need it.

'anything until end' is now shorthand for 'anything until: nothing',
which doesn't seem to affect the semantics much.  For the old semantics,
'anything until deadlock on ACTOR' still exists, but this is mainly an
oddity I need to work out.