Add OptFragments
This has been done because UML has OptFragments, and, while we can
encode them as alts with an empty second leg, the encoding is so
cumbersome that it's hard to argue against not just having the feature
directly.

This commit also refactors block generation to be uniform across all
blocks, using a header/body encoding where the body generation is
common code and the header is block-specific.