Make until fragments implicitly cold (for now)
This tries to fix a semantic oddity where 'anything until end' requires
timelock, because 'end' expands to 'SKIP' and SKIP terminates
immediately.

This is a quickfix, because I feel like there'll be at least one
situation where we want to be able to reason about the model being
forced to accept the release condition.