Add implicit SKIP_ANYTIME to end of Subsequences
This reverts a7b70d0af03451790eff900c216a5ea6fd84f71d, effectively;
instead of adding the SKIP_ANYTIME to the top of an UntilFragment we
add it to the end of the Subsequence.  Same effect in terms of making
empty subsequences accept any number of tocks (I think?); but also an
equal caveat of this being a bit of a hack.