Remove Bindings; use Variables instead
Per discussion with Ana a while ago.

There is now a variable list available in each Interaction for
collecting variables to use in erstwhile bindings.

Variable expressions are still called ConstExprs; this is a misnomer,
but we'll be removing CertExprs soon anyway.

Initial scoping work on making memory variables accessible to
expressions, but we need to do more scoping work to add types in.