Allow uniform one-liner BlockFragments
So, instead of

```
anything in X until Y
opt:
  Z
```

we now write:

```
anything in X until: Y
opt: Z
```

In other words, wherever an InteractionOperand should follow at the
end of a BlockFragment, we can use an inline syntax where there is a ':'
and one InteractionFragment.  We might add guards to this later on.