Finish bounds simplification
Per discussion with Ana, this has been renovated a bit to make it less
like RoboChart and more like UML.  In other words, the expressions are
separate, and we allow four cases (unbounded, exact bound, lower bound,
both bounds).

Currently the way to express an exact bound is to put it as the upper
bound and leave the lower bound empty.  This is a little surprising, but
I'm unsure of a better way that doesn't involve expanding out the
classes; maybe we should, I'm unsure.