Consolidate and fix up parameterisation resolution
Target parameters now carry around the container on which the constant
was discovered.  We then use the existing CGeneratorUtils definition of
constant IDs by passing through this extra information, rather than
(incorrectly) trying to generate the ID based on the object graph
(which attributed constants to interfaces instead of where they were
actually becoming parts of a parameterisation).

Also add a test to make sure that we don't pick up the wrong container.

Fixes #106.