Flip scoping of topics when we're in a ModuleTarget
This is a gotcha that RoboCert used to handle correctly then unlearned
when I cleaned up the event resolution logic: the visible surface of
events at the module/platform boundary level is the *platform* surface,
not the module one.  As such, the module ones should not be in scope,
but the platform ones should, inverting the usual logic of 'hide the
world but show the component'.

Fixes #118.