Align syntax with Holly's work
wait (X on Y) -> wait (X) on Y
deadline (X on Y) -> deadline (X) on Y

'on' can now be followed by the optional noise word 'actor'.

This aligns the syntax with Holly Hendry's sequence diagram work.

The main reason we had the previous syntax was to avoid the clash
on 'deadline (X) on Y Y->>|: op foo()' where the only thing
separating the two actors Y was a space.  This isn't a syntax
ambiguity, however, so we'll allow it from now on.